+++ app/screens/acknowledgement_screen.py
@@ -1,9 +1,7 @@
-from kivy.uix.label import Label
 from kivy.uix.screenmanager import Screen
 
 from helper.strings import Strings, Acknowledgements_Text
 from helper.functions import change_screen_to
-
 
 strings = Strings()
 acknowledgement_text = Acknowledgements_Text().TEXT
@@ -14,6 +12,7 @@ class Acknowledgment(Screen):
         super(Acknowledgment, self).__init__(**kwargs)
 
         self.ids.header_label.text = strings.BUTTON_STR_ACKNOWLEDGEMENT
+
         self.ids.acknowledgement_label.text = acknowledgement_text + "\n\n\n"
 
     def go_back(self, *args) -> None: