+++ app/screens/firetruck_exam.py
@@ -1,67 +1,49 @@
 from kivy.uix.button import Button
-from kivy.uix.label import Label
 from kivy.uix.screenmanager import Screen
 from kivy.animation import Animation
 from kivy.clock import Clock
 
-from random import shuffle
-
-from helper.functions import get_ToolQuestion_instances, change_screen_to
-from helper.file_handling import save_to_scores_file, get_score_value
-from helper.settings import Settings, Strings
+from helper.file_handling import (
+    get_selected_city_state,
+    get_selected_firetruck,
+)
+from helper.settings import Settings
+from helper.strings import Strings
 from helper.game_class import GameCore
-from helper.firetruck_layouts import build_answer_layout, build_second_answer_layout
+from helper.firetruck_layouts import build_answer_layout  # , build_second_answer_layout
+
+from screens.screen_base import BaseMethods
 
 
 settings = Settings()
 strings = Strings()
 
 
-class Firetruck_Exam(Screen):
-    def select_firetruck(self, selected_firetruck: str):
-        # troubleshooting: fix firetruck
-        # self.selected_firetruck = "Tank1" "Rüst+Lösch"
-        self.selected_firetruck = selected_firetruck
+class Firetruck_Exam(Screen, BaseMethods):
+    def on_pre_enter(self):
+        self.selected_city, _ = get_selected_city_state()
 
-        self.ids.firetruck_label.text = selected_firetruck
+        self.selected_firetruck = get_selected_firetruck()
 
-    def update_strike_label(self):
-        self.ids.strike_label.text = str(self.game.answers_correct_strike)
+        self.current_screen = self.get_current_screen()
 
-    def update_high_strike_label(self):
-        self.ids.high_strike_label.text = f"Best: {str(self.current_high_strike)}"
+        self.ids.firetruck_label.text = self.selected_firetruck
 
-    def reset_strike(self, *arg):
-        self.game.answers_correct_strike = 0
-        self.update_strike_label()
-
-    def increment_strike(self):
-        self.game.answers_correct_strike += settings.FIRETRUCK_TRAINING_CORRECT_POINTS
-        self.update_strike_label()
-
-    def reset_tool_list(self):
-        (self.firetruck_rooms, self.tool_questions) = get_ToolQuestion_instances(
-            self.selected_firetruck
-        )
-
-        shuffle(self.tool_questions)
+        # ?
 
     def play(self):
-        # init GameCore class instance and layouts
         self.game = GameCore()
+
         self.first_layout = None
         self.second_layout = None
 
-        # (re)set game specific elements
         self.reset_tool_list()
 
-        self.reset_strike()
+        self.load_high_score()
 
-        self.current_high_strike = get_score_value(
-            self.selected_firetruck, "high_strike"
-        )
+        self.reset_score()
 
-        self.update_high_strike_label()
+        self.update_score_labels()
 
         self.next_tool()
 
@@ -87,8 +69,8 @@ class Firetruck_Exam(Screen):
         # Reset image boxes
         self.ids.firetruck_rooms_layout.clear_widgets()
 
-        # troubleshooting: fix tool
-        # self.current_tool = "Handfunkgerät"  # "Druckschlauch B"
+        # troubleshooting: fix first popped tool
+        # self.set_first_tool("Unterlegplatte")  # for testing
         self.current_tool_question = self.tool_questions.pop()
 
         self.ids.tool_label.text = self.current_tool_question.tool
@@ -99,28 +81,6 @@ class Firetruck_Exam(Screen):
         self.first_layout.pos_hint = {"x": 0}
         self.ids.firetruck_rooms_layout.add_widget(self.first_layout)
 
-    def correct_answer(self):
-        self.increment_strike()
-
-        self.game.answers_correct_total += settings.FIRETRUCK_TRAINING_CORRECT_POINTS
-
-        if self.game.answers_correct_strike > self.current_high_strike:
-            self.current_high_strike = self.game.answers_correct_strike
-            self.update_high_strike_label()
-            save_to_scores_file(
-                self.selected_firetruck, "high_strike", self.game.answers_correct_strike
-            )
-
-        self.feedback_green = True
-
-    def incorrect_answer(self):
-        self.feedback_green = False
-
-        # self.reset_strike()
-        Clock.schedule_once(self.reset_strike, settings.FIRETRUCK_TRAINING_FEEDBACK_SEC)
-
-        # todo: check for PB score!
-
     def on_answer(self, instance):
         if not self.accept_answers:  # Check if answer processing is enabled
             return  # Ignore the button press if answer processing is disabled
@@ -129,46 +89,8 @@ class Firetruck_Exam(Screen):
         if instance.text in self.current_tool_question.room_answered:
             return
 
-        float_layout = self.ids.firetruck_rooms_layout.children[
-            0
-        ]  # children is reversed
-
-        # indicate if correct or incorrect answer
-        # for single correct answer
-        if len(self.current_tool_question.rooms_to_be_answered) <= 1:
-            # always identify and indicate the correct answer
-            # for child in children:
-            for child in float_layout.children:
-                if isinstance(child, Button):
-                    if child.text in self.current_tool_question.rooms:
-                        child.background_color = (0, 1, 0, 1)
-            # if, indicate incorrect answer
-            if instance.text not in self.current_tool_question.rooms:
-                instance.background_color = (1, 0, 0, 1)
-
-        # for multiple correct answers
-        else:
-            # document given answers in class instance
-            self.current_tool_question.room_answered.append(instance.text)
-
-            if instance.text not in self.current_tool_question.rooms:
-                # if, indicate incorrect and all correct answers and close
-                instance.background_color = (1, 0, 0, 1)
-                # for child in children:
-                for child in float_layout.children:
-                    if isinstance(child, Button):
-                        if child.text in self.current_tool_question.rooms:
-                            child.background_color = (0, 1, 0, 1)
-                pass
-
-            else:
-                # answer in correct answers
-                instance.background_color = (0, 0, 1, 1)
-
-                self.ids.tool_label.text += "\n"
-                self.ids.tool_label.text += strings.HINT_STR_MULTIPLE_ANSWERS
-
-                return
+        if self.color_layout(instance):
+            return
 
         # document given answers in class instance
         self.current_tool_question.room_answered.append(instance.text)
@@ -250,6 +172,3 @@ class Firetruck_Exam(Screen):
         ):
             self.ids.firetruck_rooms_layout.remove_widget(self.second_layout)
             self.second_layout = None
-
-    def go_back(self, *args) -> None:
-        change_screen_to("firetruck_menu")